+++ src/js/@shared/env.ts
