+++ src/js/model/index.ts
@@ -1,4 +1,4 @@
-import { ENV } from '../@shared/constants/env.js';
+import { ENV } from '../@shared/env.js';
 import { searchPageModel } from './search-page.js';
 import { mainPageModel } from './main-page.js';
