+++ src/js/view/index.ts
@@ -1,9 +1,9 @@
 import { YoutubeItem, YoutubeResponse } from '../model/get-query-string.js';
 import { model, VideoModel } from '../model/index.js';
 import { getModalWrapper } from './search-page.js';
-import { ENV } from '../@shared/constants/env.js';
+import { ENV } from '../@shared/env.js';
 import { renderMainPage } from './main-page.js';
-import { $ } from '../@shared/utils/utils.js';
+import { $ } from '../@shared/utils.js';
 
 export function getVideoHTML(data: VideoModel[], wrapper: Function): string {
   if (!data.length) {
