+++ src/js/view/liked-page.ts
@@ -1,7 +1,7 @@
-import { $, removeInnerHTML } from '../@shared/utils/utils.js';
 import { getToWatchVideoWrapper } from './main-page.js';
 import { model, VideoModel } from '../model/index.js';
 import { getVideoHTML } from './index.js';
+import { $ } from '../@shared/utils.js';
 
 export const renderLikedPage = () => {
   const videosWatched: VideoModel[] = (<VideoModel[]>model.getLocalStorageItem('videos')).filter((x: VideoModel) => x.checkLike === true);
